+++ app/backend/index.ts
@@ -1,0 +1,1 @@
+console.log("sweet");
